chore: move Table and UserColumn to postgres/database_tables.ts
