+++ src/postgres/database_tables.ts
@@ -1,0 +1,11 @@
+export enum Table {
+  users = 'users',
+}
+
+export enum UserColumn {
+  id = 'userid',
+  username = 'username',
+  email = 'email',
+  password = 'password',
+  created_at = 'created_at'
+}
